+++ lib/ui/screens/about/about_screen.dart
@@ -34,6 +34,7 @@ class AboutScreen extends StatelessWidget {
       appBar: AppBar(
         backgroundColor: Theme.of(context).colorScheme.primary,
         foregroundColor: Theme.of(context).colorScheme.onPrimary,
+        centerTitle: true,
         title: const Text('About'),
       ),
       body: Container(