+++ lib/ui/screens/archive/archive_screen.dart
@@ -121,6 +121,7 @@ class _ArchiveScreenState extends State<ArchiveScreen> {
       appBar: AppBar(
         backgroundColor: Theme.of(context).colorScheme.primary,
         foregroundColor: Theme.of(context).colorScheme.onPrimary,
+        centerTitle: true,
         title: Text(_currentFolderName),
         leading: IconButton(
           icon: const Icon(